0.1.6 update
