+++ PersonalTeleporter_0.1.6/graphics/hitmarker.png
