+++ PersonalTeleporter_0.1.6/graphics/tiers/orange/spritesheet.png
