+++ PersonalTeleporter_0.1.6/info.json
@@ -2,7 +2,7 @@
 
 {
   "name": "PersonalTeleporter",
-  "version": "0.1.5",
+  "version": "0.1.6",
   "title": "PersonalTeleporter",
   "author": "WhatheWorld",
   "description": "This mod adds personal Teleaporter to the game. A lot of Thanks to iUltimateLP and his mod SimpleTeleporters for inspiration and for the use of His Code and graphics",
